janky forced newlines
futile as console linesize etc still can lead to hyphenation and breaks
and no man or roff or pager options seem to clearly fix this so pandoc is helpless and
the ugly markdown is pointless too.  Suggest just a prominent link to the markdown online is the way to go.